Create block_height_compare.sh
Since all tendermint based nodes go into slow sync mode after the output of "catching_up" is "false", Jackal nodes (validators) will fall behind
This script will check the block heights of a reference node (polkachu) and compare it with local block height. It will then check if the local node is catching up or not. If it's NOT catching up AND the block height difference is more than 2, it will restart the canined service.